layaway/app/requirements.txt
bcrypt==4.0.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.2
PyJWT==2.5.0
pymongo==4.2.0
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.12
Werkzeug==2.2.2
